webapp/requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Werkzeug==0.14.1
scikit-learn==0.19.2
google-cloud-bigquery==1.22.0
pandas==0.25.2
numpy==1.17.2
Pillow==5.2.0
scipy==1.1.0
